requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
opencv-python==4.9.0.80
numpy==1.26.4
scikit-learn==1.5.0
insightface==0.7.3
onnxruntime==1.18.0
apscheduler==3.10.4
pillow==10.3.0